LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/cors-crossorigin-requests-expected.txt

Blocked access to external URL http://www2.localhost:8800/html/semantics/scripting-1/the-script-element/json-module/data.json
Blocked access to external URL http://www2.localhost:8800/html/semantics/scripting-1/the-script-element/json-module/data.json?pipe=header(Access-Control-Allow-Origin,*)
Blocked access to external URL http://www2.localhost:8800/html/semantics/scripting-1/the-script-element/json-module/parse-error.json?pipe=header(Access-Control-Allow-Origin,*)
json-module-crossorigin



FAIL Imported JSON module, cross-origin with CORS assert_equals: Unexpected _log value expected "imported JSON: 42" but got ""
PASS Imported JSON module, cross-origin, missing CORS ACAO header
FAIL Imported JSON module with parse error, cross-origin, with CORS assert_equals: Unexpected _log value expected "0-0" but got ""
